refactor(primitives): extract draw_primitives.rs from pageable.rs (PR 8j Task 1)
Move geometry types (Pt, Size, Affine2D, Point2, Rect, Quad, BreakBefore/After/Inside),
collectors (DestinationRegistry, LinkCollector, LinkOccurrence), Canvas + draw_with_opacity,
style types (BoxShadow, BlockStyle, BorderStyleValue, Overflow, Bg* enums, BackgroundLayer),
drawing helpers (build_rounded_rect_path, compute_overflow_clip_path, draw_block_border, etc.),
BookmarkEntry + BookmarkCollector, clamp_marker_size, and matrix_test_util into the new
draw_primitives.rs module. pageable.rs re-exports all via `pub use crate::draw_primitives::*`.
