+++ crates/fulgur/src/lib.rs
@@ -20,6 +20,7 @@ pub(crate) mod link;
 pub mod multicol_layout;
 pub(crate) mod net;
 pub mod outline;
+pub mod draw_primitives;
 pub mod pageable;
 #[doc(hidden)]
 pub mod pagination_layout;
